backoffice added

+++ SocialDevelop/src/java/socialdevelop/controller/rmSkillBack.java
@@ -18,6 +18,8 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
+import socialdevelop.data.model.Admin;
+import socialdevelop.data.model.Developer;
 import socialdevelop.data.model.Skill;
 import socialdevelop.data.model.SocialDevelopDataLayer;
 
@@ -43,24 +45,33 @@ public class rmSkillBack extends SocialDevelopBaseController {
                 String u = (String) s.getAttribute("previous_url");
                 if(s.getAttribute("userid") != null && ((int) s.getAttribute("userid"))>0) {
                     if(s.getAttribute("previous_url") != null && ((String) s.getAttribute("previous_url")).equals("/socialdevelop/BackEndSkill")){
+                        
             
                         SocialDevelopDataLayer datalayer = (SocialDevelopDataLayer) request.getAttribute("datalayer");
+                        Developer dev = datalayer.getDeveloper((int) s.getAttribute("userid"));
+                        Admin admin = datalayer.getAdmin(dev.getKey());
+                        if (admin != null && admin.getDevelperKey() > 0){
                         
-                        
-                        Skill deadSkill = datalayer.getSkill(parseInt(request.getParameter("rm-skill-b")));
-                        List <Skill> skills = deadSkill.getChild();
-                        
-                        if(!(skills.isEmpty())){
-                            for (Skill skill : skills){
-                                skill.setParentKey(NULL);
-                                datalayer.storeSkill(skill);
-                            }
-                        }
-                        datalayer.deleteSkill(deadSkill);
+                            Skill deadSkill = datalayer.getSkill(parseInt(request.getParameter("rm-skill-b")));
+                            List <Skill> skills = deadSkill.getChild();
 
-                        datalayer.destroy();
+                            if(!(skills.isEmpty())){
+                                for (Skill skill : skills){
+                                    skill.setParentKey(NULL);
+                                    datalayer.storeSkill(skill);
+                                }
+                            }
+                            datalayer.deleteSkill(deadSkill);
+
+                            datalayer.destroy();
+                            s.removeAttribute("previous_url");
+                            response.sendRedirect(u.split("/")[2]);
+                        
+                        }else{
                         s.removeAttribute("previous_url");
-                        response.sendRedirect(u.split("/")[2]);
+                        response.sendRedirect("index");
+
+                    }
 
                     }else{
                         s.removeAttribute("previous_url");
