+++ SocialDevelop/src/java/socialdevelop/controller/rmType.java
@@ -1,8 +1,8 @@
 /*
-* To change this license header, choose License Headers in Project Properties.
-* To change this template file, choose Tools | Templates
-* and open the template in the editor.
-*/
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
 package socialdevelop.controller;
 
 import it.univaq.f4i.iw.framework.result.FailureResult;
@@ -17,6 +17,8 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
+import socialdevelop.data.model.Admin;
+import socialdevelop.data.model.Developer;
 import socialdevelop.data.model.Skill;
 import socialdevelop.data.model.SocialDevelopDataLayer;
 import socialdevelop.data.model.Type;
@@ -32,57 +34,72 @@ public class rmType extends SocialDevelopBaseController {
             (new FailureResult(getServletContext())).activate((Exception) request.getAttribute("exception"), request, response);
         }
     }
-    
-    private void rm_type(HttpServletRequest request, HttpServletResponse response) throws SQLException, IOException, NamingException, NoSuchAlgorithmException, Exception {
-        HttpSession s = request.getSession(true);
-        if(s.getAttribute("userid") != null && ((int) s.getAttribute("userid"))>0) {
-            
-            SocialDevelopDataLayer datalayer = (SocialDevelopDataLayer) request.getAttribute("datalayer");     
-            Type type = datalayer.getType(datalayer.getTypeByName(request.getParameter("type")));       
-            List <Skill> skills =  datalayer.getSkillsByType(type.getKey()); 
-            if(!(skills.isEmpty())){
-                for( Skill skill : skills) {
-                    
-                    skill.setType_key(NULL);
-                    datalayer.storeSkill(skill);
-                }
-            }
-            int ret = datalayer.deleteType(type);
-            datalayer.destroy();
-            response.setContentType("text/plain");
-            response.setCharacterEncoding("UTF-8");
-            PrintWriter out = response.getWriter();
-            
-            try {
-                out.println(ret);
-            } finally {
-                out.close();
-            }
-        }else{
-            s.removeAttribute("previous_url");
-            response.sendRedirect("index");
-        }
         
-    }
+     private void rm_type(HttpServletRequest request, HttpServletResponse response) throws SQLException, IOException, NamingException, NoSuchAlgorithmException, Exception {
+        HttpSession s = request.getSession(true);
+            if(s.getAttribute("userid") != null && ((int) s.getAttribute("userid"))>0) {
+                
+                SocialDevelopDataLayer datalayer = (SocialDevelopDataLayer) request.getAttribute("datalayer");
+                Developer dev = datalayer.getDeveloper((int) s.getAttribute("userid"));
+                Admin admin = datalayer.getAdmin(dev.getKey());
+                if (admin != null && admin.getDevelperKey() > 0){
+
+                    Type type = datalayer.getType(datalayer.getTypeByName(request.getParameter("type")));
+
+                    List <Skill> skills =  datalayer.getSkillsByType(type.getKey());
+
+                    if(!(skills.isEmpty())){
+
+                        for( Skill skill : skills) {
+
+                            skill.setType_key(NULL);
+                            datalayer.storeSkill(skill);
+                        }
+                    }
+
+                    int ret = datalayer.deleteType(type);
+
+
+
+                    datalayer.destroy();
+                    response.setContentType("text/plain"); 
+                    response.setCharacterEncoding("UTF-8"); 
+                    PrintWriter out = response.getWriter();
+
+                    try {
+                        out.println(ret);
+                    } finally {
+                        out.close();
+                    }
+                }else{
+                s.removeAttribute("previous_url");
+                response.sendRedirect("index");
+            }
+            }else{
+                s.removeAttribute("previous_url");
+                response.sendRedirect("index");
+            }
     
+     }
     
+     
     @Override
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException {
         try {
             rm_type(request, response);
         } catch (SQLException ex) {
-            request.setAttribute("exception", ex);
-            action_error(request, response);
+           request.setAttribute("exception", ex);
+            action_error(request, response);  
         } catch (NamingException ex) {
-            request.setAttribute("exception", ex);
-            action_error(request, response);
-        } catch (Exception ex) {
-            request.setAttribute("exception", ex);
-            action_error(request, response);
-        }
+           request.setAttribute("exception", ex);
+            action_error(request, response);  
+        } catch (Exception ex) { 
+           request.setAttribute("exception", ex);
+            action_error(request, response);  
+        } 
     }
-    
+
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
     
     public String getServletInfo() {
